crud categories ok

+++ src/app/modules/folders/components/folder/folder.component.ts
@@ -5,6 +5,9 @@ import { DocumentsQuery } from '../../../../akita/query/documents.query';
 import { DocumentService } from '../../../../akita/service/documents.service';
 import { MenuController } from '@ionic/angular';
 import { UIService } from '../../../../services/ui.service';
+import { CategoriesServices } from '../../categories.services';
+import { AuthenticationService } from '../../../../services/authentication.service';
+import { first } from 'rxjs/operators';
 
 @Component({
   selector: 'app-folder',
@@ -18,7 +21,9 @@ export class FolderComponent implements OnInit {
     private documentQuery: DocumentsQuery,
     private documentService: DocumentService,
     private menuController: MenuController,
-    private ui: UIService) { }
+    private ui: UIService,
+    private catService: CategoriesServices,
+    private authService: AuthenticationService) { }
 
   ngOnInit() {
     this.documentQuery.getDocumentsByFolder(this.folder.id).subscribe(data=>{
@@ -31,14 +36,31 @@ export class FolderComponent implements OnInit {
     this.ui.presentAlertConfirm("Al borrarla se eliminarán todos los archivos relacionados","¿Desea eliminar la categoria?",
     (ok) => {
  
-      
-      let otros = this.documentQuery.getDocumentsByFolderObject(id);
-      //console.log(otros);
-      otros.forEach(item=>{
-        this.documentService.delete(item.id);
-      })      
-      this.service.delete(id);    
-      this.documentService.updateFolder(0);
+      this.catService
+      .delete(id, this.authService.currentOwnerValue.id)
+      .pipe(first())
+      .subscribe({
+        next: (res) => {   
+          if(res.result){
+            let otros = this.documentQuery.getDocumentsByFolderObject(id);
+            //console.log(otros);
+            otros.forEach(item=>{
+              this.documentService.delete(item.id);
+            })      
+            this.service.delete(id);    
+            this.documentService.updateFolder(0);
+          } else {
+            this.ui.presentToast("Se encontro un errro al eliminar la categoria")
+          }
+         
+         
+          
+        },
+        error: (error) => {
+          
+        },
+      });
+    
     }, 
     (error) => console.log(error))
     ;
